refac(prediction writer): rename utils.py to caches.py
