+++ src/careamics/lightning/callbacks/prediction_writer_callback/write_strategy/write_image.py
@@ -11,7 +11,7 @@ from torch.utils.data import DataLoader
 from careamics.dataset import IterablePredDataset
 from careamics.file_io import WriteFunc
 
-from .utils import SampleCache
+from .caches import SampleCache
 
 
 class WriteImage:
